backend/requirements.txt
flask
flask-cors
psycopg2-binary
sqlalchemy
python-dotenv
flask-sqlalchemy
marshmallow
flask-marshmallow
